Allow exception nesting
Allow exceptions to be nested, which is especially useful with urgent
interrupts while processing an exception.

The implementation counts up the nesting level with each call to an
exception. In the outer exception (level 1), the exception stack is
started. All nested exceptions just reserve the redzone (scratch
memory that may be used by compiler) and exception context on the
stack, but then process on the same scratch.

Restriction:

 * Impure pointers are shared among all exceptions. This may be solved
   by creating an impure data structure in the stack frame with each
   nested exception.